Type the node paged-envelope counters as integer (OpenAPI parity)
The conformance parity gate compares each backend's OpenAPI property
types against node.  The paged envelope's `page`/`pageSize`/`total`/
`totalPages` are integers on every other backend (.NET/Java/Python/
Phoenix emit `type: integer`), but node's zod schema used a bare
`z.number()` → `type: number`, drifting the property-type check for
BuildPaged/EngineerPaged/ProjectPaged/SquadPaged (node=number,
dotnet=integer).

Mark them `z.number().int()` (→ `integer`) at the single backend site
(Hono routes-builder, shared by node@v4/v5) and the two frontend
api-client schemas (shared React/Vue + Svelte) so the client parser
matches the wire contract.  Integer `format` differences are already
ignored by the parity normalizer (node's plain-int fields carry no
`format` and pass against dotnet's `int32` today), so no format drift.

Verified end-to-end against live-booted backends: node now emits
`{type: integer}` + all-required for every *Paged schema, matching the
.NET spec.  Re-baselined the acme baseline-output fixtures (the paged
schema lines + the .NET RequiredFromCtorParamFilter Paged branch) and
updated the two paged-emit unit-test assertions.

+++ src/generator/_frontend/api-module.ts
@@ -215,7 +215,10 @@ export function buildApiModule(
       if (!paged || pagedSeen.has(paged.name)) continue;
       pagedSeen.add(paged.name);
       lines.push(
-        `export const ${paged.name} = z.object({ items: z.array(${zodForResponseInner(paged.arg)}), page: z.number(), pageSize: z.number(), total: z.number(), totalPages: z.number() });`,
+        // Integer pagination counters — match the backend wire contract
+        // (every backend's OpenAPI types these `integer`; see the Hono
+        // routes-builder paged schema).
+        `export const ${paged.name} = z.object({ items: z.array(${zodForResponseInner(paged.arg)}), page: z.number().int(), pageSize: z.number().int(), total: z.number().int(), totalPages: z.number().int() });`,
       );
       lines.push(`export type ${paged.name} = z.infer<typeof ${paged.name}>;`);
     }
