+++ src/platform/hono/v4/routes-builder.ts
@@ -393,7 +393,11 @@ export function buildRoutesFile(
       if (!paged || pagedSeen.has(paged.name)) continue;
       pagedSeen.add(paged.name);
       lines.push(
-        `export const ${paged.name} = z.object({ items: z.array(${zodForResponse(paged.arg, false)}), page: z.number(), pageSize: z.number(), total: z.number(), totalPages: z.number() }).openapi("${paged.name}");`,
+        // The pagination counters are integers on every other backend's
+        // OpenAPI (.NET/Java/Python/Phoenix emit `integer`), so mark them
+        // `z.number().int()` here too — a bare `z.number()` emits `number` and
+        // drifts the conformance-parity property-type check (M-T2.6).
+        `export const ${paged.name} = z.object({ items: z.array(${zodForResponse(paged.arg, false)}), page: z.number().int(), pageSize: z.number().int(), total: z.number().int(), totalPages: z.number().int() }).openapi("${paged.name}");`,
       );
     }
   }
